// DXC_Frontend/src/api/models/iconDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface IconDto {
  publicId?: string
  /** @nullable */
  iconCategoryPublicId?: string | null
  /** @nullable */
  iconGroupPublicId?: string | null
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string
  /**
   * @maxLength 500
   * @nullable
   */
  description?: string | null
  /**
   * @maxLength 500
   * @nullable
   */
  iconImageUrl?: string | null
  /** @nullable */
  imagePublicId?: string | null
  /**
   * @minLength 1
   * @maxLength 20
   */
  iconType: string
  /**
   * @maxLength 1000
   * @nullable
   */
  screenParams?: string | null
  /**
   * @maxLength 500
   * @nullable
   */
  webLink?: string | null
  /** @nullable */
  linkAndroid?: string | null
  /** @nullable */
  linkIOS?: string | null
  displayOrder?: number
  thuTu?: number
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
  /** @nullable */
  iconCategoryName?: string | null
  /** @nullable */
  iconGroupName?: string | null
}